requirements-dev.txt
# Development dependencies for DataSON Benchmarks
# These are needed for workflow generation, testing, and development

# Workflow generation
ruamel.yaml>=0.18.0

# Code quality
ruff>=0.6.0
pre-commit>=3.0.0

# Testing  
pytest>=7.0.0
pytest-cov>=4.0.0

# Schema validation (used by pre-commit)
check-jsonschema>=0.28.0